Create COVID19 Vulnerability
This project is part of a group project for the course that I took for my M.Sc Economics program at LMU Munich. The course is called, Machine Learning: Applications in Economic Research. In this project, we are required to look for real data and apply ML methods we learned in class. The topic we chose is the vulnerability of countries against the COVID19 shock. My contribution of the project was mainly on the coding part where I prepared the data, built OLS and Random Forests models. Then, I compared the performances of these models side-by-side and later, compared them to the third model, LASSO Regression coded by my groupmate.